refactor: polish test code
