+++ .github/workflows/test.yml
@@ -3,8 +3,18 @@ name: Robot CPP Tests
 on:
   push:
     branches: [ master ]
+    paths:
+      - 'src/**'
+      - 'tests/**'
+      - '.github/**'
+      - 'CMakeLists.txt'
   pull_request:
     branches: [ master ]
+    paths:
+      - 'src/**'
+      - 'tests/**'
+      - '.github/**'
+      - 'CMakeLists.txt'
 
 jobs:
   test-macos:
