+++ CMakeLists.txt
@@ -2,8 +2,7 @@ cmake_minimum_required(VERSION 3.21)
 
 project(RobotCPP)
 
-# Set modern C++ standard for all targets
-set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD 23)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
