+++ README.md
@@ -1,6 +1,6 @@
 # Robot CPP
 
-![master](https://github.com/developer239/robot-cpp/actions/workflows/ci.yml/badge.svg)
+[![Robot CPP Tests](https://github.com/developer239/robot-cpp/actions/workflows/test.yml/badge.svg)](https://github.com/developer239/robot-cpp/actions/workflows/test.yml)
 ![Windows](https://img.shields.io/badge/Windows-0078D6?style=for-the-badge&logo=windows&logoColor=white)
 ![macOS](https://img.shields.io/badge/mac%20os-000000?style=for-the-badge&logo=macos&logoColor=F0F0F0)
