--- tests/README.md
@@ -1,13 +1,0 @@
-# Test Assets Directory
-
-This directory contains assets required for testing the Robot CPP library.
-
-## Structure
-
-- `expected/` - Contains reference images for comparison in screen capture tests
-- `temp/` - Temporary directory for test outputs (screenshots, logs, etc.)
-
-## Usage
-
-The SDL test application will save screenshots to this directory during tests.
-When running tests, you can examine these screenshots to verify visual output.
